# api-docs/openAPI-specification.yml
openapi: 3.0.1
info:
  version: "1.0.0"
  title: RUMbus API
  description: >-
    RUMbus is a system to provide real-time information on the location of
    buses, as well as estimated arrival times at stops for the transportation
    service of the University of Puerto Rico at Mayaguez (UPRM).
...
paths:
...
  /drivers/search:
    get:
      tags:
        - Drivers
      summary: Search for drivers with the specified parameters
      parameters:
        - name: license
          in: query
          schema:
            type: string
          description: Driver license number
        - name: id
          in: query
          schema:
            type: string
          description: Driver id
        - name: full_name
          in: query
          schema:
            type: string
          description: Driver full name
        - name: skip
          in: query
          schema:
            type: string
          description: Parameter for pagination
        - name: limit
          in: query
          schema:
            type: string
          description: Parameter for pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                    full_name:
                      type: string
                    license:
                      type: integer
                    created_on:
                      type: string
                    updated_on:
                      type: string
              examples:
                '0':
                  value: >-
                    [{"id":"1","full_name":"Test
                    John","license":54321,"updated_on":"2020-01-21T00:27:10.000Z","created_on":"2020-01-21T00:27:10.000Z"}]
                '1':
                  value: []
        '500':
          $ref: '#/components/responses/500'
  /drivers/{driver_id}:
    get:
      tags:
        - Drivers
      summary: Get driver by its id
      parameters:
        - in: path
          name: driver_id
          schema:
            type: string
          required: true
          description: ID of the driver to fetch
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  license:
                    type: integer
                  full_name:
                    type: string
                  updated_on:
                    type: string
                  created_on:
                    type: string
              examples:
                '0':
                  value: >-
                    {"id":"1","full_name":"Test
                    John","license":54321,"updated_on":"2020-01-21T00:27:10.000Z","created_on":"2020-01-21T00:27:10.000Z"}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: '{"code":"NotFound","message":"Driver with id: abc404 not found"}'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Drivers
      summary: Update driver by its id
      parameters:
        - in: path
          name: driver_id
          schema:
            type: string
          required: true
          description: ID of the driver to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                license:
                  type: integer
            examples:
              '0':
                value: "{ \r\n    \"license\": 123456\r\n}"
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'

    delete:
      tags:
        - Drivers
      summary: Delete driver by its id
      parameters:
        - in: path
          name: driver_id
          schema:
            type: string
          required: true
          description: ID of the driver to delete
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '204':
          description: No Content
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
          examples:
            'Server Error':
              value: '{"code":"InternalServer","message":"An internal server error ocurred"}'